[F] Expose force overwrite in entity store
In some cases, responses from the API will return a partial
representation of a model for performance reasons. By default,
the entity store reducer will not overwrite an entity in the store
that has been fully serialized with a partially serialzied
representation of the same model. This behavior was implemented to
prevent an API request from one part of the UI that included a
partially serialized relationship interfering with the full
representation in another part of the UI. There are very few cases
where this can happen, but we did run into one in the reader.

That said, we're now seeing the opposite problem, in which a
partially serialied entity needs to replace a fully serialized
entity. For example, when a user is managing project texts in the
backend for a project, we use the partially serialized texts in
the Project API response to populate the interface. When a user
changes the position of a text, we re-fetch the project and the
entites update. However, if the user edits a text, that triggers
the loading of the fully serialized text. When they go back to
sort the texts, the partially serialized texts coming back in
Project.relationships.texts are unable to overwrite the fully
serialized, cached text in the store.

To address these edge cases, it is now possible to pass a "force"
option in the request options hash. We may nede to revisit this
full/partial distinction in the future.