[C] Rename Collection model to ResourceCollection
This commit renames and refactors uses of Collection as ResourceCollection.
We're doing this because it is no longer immediately clear what collection
refers to.  With the addition of ProjectCollections and potentially future
collection kinds, we should take care of this now, rather than later as
things get more complicated.

This particular commit is limited to the change in the API, which will
facilitate a refactor of the frontend components and organization.