[B] Send base model props in <Form> create request
When we have an edit <Form>, we generally pass the model being edited
as the "model" prop on the Form component. When the Form opens an
editing session, it will use that model as the "source" in the session.
When passing field values to the child form inputs, the Form container
consults the dirty model for a value, and then falls back to the source
model. With updates, the form will only post the changed (dirty) values
to the API.

In the case of a create, however, the form should post the dirty values
overlaid on the source values, instead of only posting the changed
values. With this change, we can pass scaffolded models with default
values to a "new" model form, and the form will use those values to
both populate field values (current behavior), and in the post to the
create API endpoint (updated behavior).