[B] Memoize pending action callout
Due to https://github.com/ReactTraining/react-router/issues/5099, our
connected components update more often than they should. This is
usually not a problem, but in this case it causes a new object to be
pashed to the form. This fixes that by memoizing the object and tying
the memoization to history state.