[F] Make reader iframes responsive; support H5P embeds
This commit aims to improve how inline frames (<iframe>) are rendered
in the reader. Going forward, ingested sources should set a target width
and height in pixels on all iframes. These can be set as either tag
attributes or inline CSS properties. For example, both of the following
markup examples would be supported by Manifold:

```
<iframe width="400" height="300"></iframe>
<iframe style="width: 400px; height: 300px"></iframe>
```

When the frame is rendered in the reader, Manifold will calculate the
aspect ratio based on the specified dimensions, and will render the
frame in a responsive wrapper that will adjust to the full width of the
reader.

In addition, this revision lays the groundwork for supporting specific
kinds of frames via a custom "data-manifold-iframe" attribute. If this
attribute is present on ingested sources, and if its value is a type of
iframe that Manifold recognizes, Manifold will render the frame with
additional features needed for that type of frame.

This commit introduces support for frames from H5P. H5P iframes require
a script from the H5P CDN to handle dynamic resizing based on the
contents. Manifold does not, however, allow ingested texts to include
script tags. Adding the `data-manifold-iframe` attribute with a value of
"h5p" to the iframe tells Manifold to include the H5P resizing code from
the external location. For example:

```
<iframe
  data-manifold-iframe="h5p"
  src="h5p.com/foo"
  width="1200"
  height="3000">
</iframe>
```

This is intended to be a proof of concept, and in the future we will
likely want to support other common inline frames.