[B] Fix action callout ordering
This commit fixes action callout reordering.  Previously when `addToSlot`
was run, the state update from `removeFromSlot` might not have completed.
This lead to instances (noticed with keyboard navigation) where an item
would appear in a list twice: once at its original position and once at
its new position. This change adds a callback to `removeFromSlot` that
receives the removed callout.  That callback function is then passed into
`replaceSlotInState`, which updates state and then fires the callback.

This ensures that the item has been removed from the slot before it is
added to a new position/another slot.