[B] Fix OAuth sessions not setting cookie
This commit refactors the auth cookie setting into its own function.
This function responds to the `LOGIN_SET_AUTH_TOKEN` action, which sets
the auth token in state.  With this change, the auth token will also be
set in a cookie to persist login sessions.  This will fix an issue where
logging in with OAuth would not set an auth token cookie and thus
would log users out on a refresh.

Fixes #1898