[F] Implement #renderable? on content blocks
This commit removes block specific model requirements.  Instead,
blocks can now be saved in an "incomplete" or "unrenderable" state.
This change is the first step towards implementing project templates
where blocks will be added to a project before being configured.

When a block is serialized we include whether or not it has been
configured with the minimum attributes necessary to render it in
the client.  As part of this, there has been some additional
validation added to the content block model.  This allows us to tell
a user exactly what needs to be configured for the block to become
renderable.