[B] Remove tag eager loading from projects
This was causing an issue with deleting projects were resource tags
were destroyed early.  Then when ActsAsTaggableOn went to clean up
the unused tags, it would throw an error trying to find the destroyed
tags.

Removing this eager loading does not seem to have any negative impact
on the AR time of projects controller requests. If anything, they seem
to be a hair quicker.

Fixes #1803