[B] Fix error serializing search results
This is a an interim solution to fixing an issue where serializing
search result objects.  It seems like the removal of SearchableNodes
is being interupted when a project or text is destroyed.  This leaves
orphaned SearchableNodes that get indexed by searchkick.

This commit makes the SearchResultSerializer and corresponding client
component a little more resilient to a result without an existing,
associated model in the DB.

Relates to #1958