[B] Fix error serializing search results
This commit improves search results by filtering out ones where the
associated model no longer exists.  Due to how we handle some project child
association deletions, there are cases where our search indices reference
models that no longer exist in the DB.  This then throws an error when serializing
the result and trying to render the result on the frontend.

The raw Search::Result has to be the initial point of filtering because
of pagination.  We could render null in the client and serialize results with
empty parents, but then the list of visible results would differ from the
pagination shown.

Fixes to #1958