[F] Add service to replace files that fail to process
We have had several issues in the past where Shrine attachments failed
to process and thus got stuck in the :cache storage.  At this point, the
original files are most frequently lost and the attachment is in an
irrecoverable state.

This service takes a replacement dir with files that have names matching
the stalled files.  It then attempts to find all attachments that are still
in the cache with a matching filename and replace them with the new version.

This should allow us to replace attachments after we address bugs that prevent
attachments from processing.