[R] Refactor entity list search state management
There is further room for improvement here. I’m not 100% happy with the
interface in and out of the HOC. That said, this at least moves a lot of
the logic out of the containers and into the HOC, the search component,
and the filter defintions, which is an improvement over what we have now.

Fixes #2085