[R] Replace labelId helper with react-uid
This commit fixes two issues with how unique IDs were being generated
to associate HTML form elements with their corresponding label or
"describedby" element. Use of `labelId` as a `defaultProp` was
generating the same ID for all rendered instances of a component on a
given page. In other places, use of `labelId` in the render method
was generating a different ID on every re-render, which wasn't
necessarily a problem in this context, but wasn't ideal. Both of these
issues are resolved by wrapping form elements in react-uid's renderless
`UID` component, which creates a unique ID for each instance that
doesn't change on re-render.

This commit also adds a mock for react-uid to the test suite, and
updates test snapshots.