[B] Remove rendundant SR close buttons from drawers
This commit removes extraneous screen-reader-only buttons from drawers.
First, it removes the button functionality attached to the overlay
behind the global drawer component, which is made redundant by
`focus-trap-react`'s `clickOutsideDeactivates` property. Second, it
revises the conditions for rendering a SR button in the header of
the drawer, as implemented in 81a3155. Rather than render it if
`props.includeDrawerFrontMatter` is false, it's now tied to a new prop,
`includeSRCloseButton`. This allows inclusion of a close button in the
reader TOC button without rendering unnecessary buttons in other
drawers such as reader notes.

Resolves #2241