[F] Further relax user stylesheet validation
When we extract inline styles from ingested texts we move them to class-
based selectors. We do this to make them easier to override from
stylesheets, however, this has the opposite effect in the case of google
docs. This commit wraps all the text section content in an ID which can
be used by user stylesheets to scope styles with more specificity than
the class based selectors. In addition, this commit allows user
stylesheets to include !important declarations.