[C] Improve project placeholder content
Fixes #2237

Prior to this change, unauthenticated users were seeing placeholders
that invited them to create content. This revision improves the place-
holder components to provide more appropriate messages to users who
can't create projects.

To generally improve the experience on instances that have no visible
projects to the current user, we needed to expand the section of the
settings object that includes calculated properties, such as whether
or not the instance has projects. To perform these calculations, the
settings object needs to be aware of the logged in user. This
requirement led to changes in how we bootstrap the settings, and led
us to expose an authToken option in the API request hash.

If the authToken is explicitly included in the request options, the
middleware will use it. Generally speaking we pull the authToken from
the authenticated user in the store. However, when we load the settings
during the server-side bootstrap, it's easier to just include the
token from the cookie. This condition supports that
possibility.