[B] Do not authenticate user until login complete
Prior to this change, the "authenticated" flag in the store was being set to
true before the auth token was stored. This change helps components to reliably
watch for a change to the authenticated boolean as a way of knowing when users
login. Prior to this change, when an unauthenticated user joined a reading
group, the user would be redirected to the reading groups view and the reading
group would be fetched before the login was complete, which would lead to an
unauthorized request.