[F] Replace Active Model Serializers
Active Model Serializers has grown a little long in the tooth and the
time has come to replace it with a faster solution. Enter fast_jsonapi.

The fast_jsonapi library was started by an employee at Netflix, and has
now been forked to a community repo here:

https://github.com/fast-jsonapi/fast_jsonapi

We've made our own fork under ManifoldScholar to support our tendency
to partially serialize some objects for performance reasons.

The motivation for replacing AMS was that we were starting to see poor
performance when large collections of resources were serialized. AMS
tries to solve a lot of problems, and a more focused lib that only
addresses the challenges involved with supporting the json-api spec
gives us a 20-30% performance gain during serialization.

Fixes #2443