[F] Improve performance of lists with many texts
When a project has hundreds of texts, the rendering cost of the project
show route is extremely high. This change, combined with the movement
away from ActiveModelSerializers, is meant to improve performance in cases.

Texts, in particular, are related to a number of database records. This
commit adds a new TextSummary model that is backed by a database view and
includes the minimal amount of data needed to render a text in a list.
We shave a second or two off of time in the database, and we simplify the
production of the json object with this change.

Fixes #2443