[F] Improve performance of lists with many projects
Similar to the text refactor, this change introduces the ProjectSummary,
which is a model backed by a DB view that includes the minimal amount of
data needed to render a project in a list. This model should always be
serializable by the project serializer when not in "full" serialization
mode.

Fixes #2443