[F] Cache formatted attributes in DB
In long lists of resources, returning cached formatted data from Redis
can be unnecessarily time consuming. For example, in a list of 600
texts, we would need to fetch the subtitle, title, and description from
the Redis cache. Each of those has a formatted and a plaintext value,
so we're looking at 600 * 3 * 2 or 3,600 Redis requests to render the
list. With this commit, the formatted attributes concern will look for
cached_<attr_name>_formatted= and cached_<attr_name>_plaintext= methods
on the model, which will be present if corresponding DB fields exist.
If the model responds to those methods, it will _also_ cache the values
in the database. The DB cache allows serializers (such as the text and
project serializers) to use the cached value instead of reaching out
to Redis.

This feature will be utilized in a subsequent commit.

Fixes #2443