[B] Remove pagination from BE PC containers
This commit represents a significant refactor to how we fetch data in
the backend project collection views. Prior to this refactor, the
various containers were dealing with CollectionProjects (CP) as a
relationship on the ProjectCollection (PC), as well as the PC's
Projects relationship, which was a HMT on CPs. To add complexity, the
CP relationship on PCs is paginated, which makes sense when users are
viewing projects on the frontend, but not when they're managing
projects in the PC on the backend.

This change builds on work done in the AMS to Fast JSON API replacement
commit to add new PC/Relationships/CP endpoints. It utilized the
ProjectSummary view so that we can, in theory, render hundreds of
projects in a PC without too big of a performance hit.

It also aims to simplify and DRY up the data fetching logic among the
various backend PC components.

Fixes #2424