[B] Restore height to annotation popup wrapper
When the panel is positioned absolutely, the .annotation-popup ancestor
element does not have height. The popup positioner needs to be able to
accurately determine the height off the .annotation-popup container so
that it can determine whether to position it above or below the
selection. One of the jobs of the popup positioner is to prevent the
popup from colliding with the edges of the window, and it cannot do this
without an accurate height measurement.

Fixes #2469