[B] Persist dirty form values when model updates
Prior to this change, any change to a model on a form edit session
would re-open the editing session and the user would lose any unsaved
changes to the model. With this change, the form will attempt to
refresh the existing session rather than opening a new editing session.
The refresh action will update the underyling model for the session and
will loop through any dirty values and see if they are still dirty by
comparing them to the model. Any values that are still dirty will be
persisted into the new editing session.

Fixes #2449