[B] Improve entityEditor refresh state change
In 9e18e8, we adjusted the entityEditor reducer to keep dirty values
in the session that don't match the source model when a new version
of the entity is sent to the form. This was done to address the issue
in #2449. This fix, however, led to a problem in saving settings. The
settings model contains nested hashes (eg, theme, general, etc.). When
we saved the model, the reducer would see the hash as changed because
it wasn't doing a deep compare. We can fix this by doing a more
thorough compare of the source and dirty objects, but for now we will
simply delete any dirty attributes that are object-like. This approach
keeps the reducer simple while still solving the problem in #2449.