[B] Adjust Notation.Marker click behavior
This narrows the conditions under which a click event makes a resource
notation active in the reader (versus linking through to view that
resource). Currently we show two different views for notations: on
smaller devices, the active one is fixed to the bottom of the viewport.
On wider devices, all notations are shown as thumbnails in the margin
of the reader text block. The previous click handler method tested
for whether the user's device had touch support, and assumed that if it
did, then that resource should be activated at the bottom of the
viewport. This conflated device size and touch support, however, and
led to an issue on laptops with touch screens, where the user could see
the active thumbnail present in the margin, but couldn't click on the
notation marker. To solve this, I've narrowed the conditions on the
click handler to only disallow marker clickthroughs when the device has
touch support AND is below the breakpoint when the thumbnails are hidden
from the margin (1235px).

Resolves #2410