[B] Skip some invalid ingestion sources
When ingesting, we can safely skip ingestion sources that are invalid
and of the type “publication_resource”. If we dont skip these, all it
takes is an errant DS_Store file to blow up a manifest ingestion, which
will try to include every file it finds in the archive, whether or not
its referenced in the manifest. This is less of a problem with EPUBs,
which require an explicit reference to every file in the package manifest.

Fixes #2612