[F] Add login page & improve logout redirection
With the feature described in #2652, we needed to rethink how we handle
login and logout. If library pages are disabled, we're no longer able to
redirect the user to the homepage on logout. If we do, we get into a
situation where the user could disable library mode in the backend, then
logout and get redirected to the home page, at which point the
RedirectIfLibraryModeDisabled component would redirect the user to an
external URL. The user would go to /backend to change the setting, only
to be redirected to /#signup, which wouldn't work.

To solve this, we need to implement a proper standalone /login route. With
this new route, after a user fails to access the backend we can redirect her
to /login, where she can login and be redirected back to the backend.

We also need to keep the user on the current page on logout to support
disabled library mode. That way, if a user logs out on a project page, she
stays on that page.