[B] Add poster fallback to interactive slides
The logic for when an iframe is to be displayed now exists in the
SlideInteractive component. A Hero simply needs to display the
iframe, so the SlideInteractive component was replaces with a
resource player. Some slight SCSS tweaks were made to shift
some of the iframe sizing from JavaScript to CSS.