[B] Improve collection request performance
Over time and with the addition of entitlements our authorizers have
become more complex and resource intensive. This is especially true for
projects and project children which delegate many ability checks to the
parent project. Prior to this commit, serializing a collection of
projects was very resource intensive, as each project included a full
ability hash. With this commit, we will no longer serialize abilties
for responses to #index actions. This could cause some regressions, and
there may be cases in the future where we want to determine whether a
user can update a resource  in a collection response. If we do end up
needing that, we’ll have to try to express the logic in SQL rather than
code so as to keep our collection routes performant.