[B] Fix missing buttons for annotation and comments
Resolves #2819.

Includes the full serialized attributes for the annotation and comment
so the UI knows that the user has permission to update and delete a
comment or annotation they created. This fixes a regression introduced
in cce22253eb1559f3bec496960bc3bab9af75d0f2 when abilities were
simplified for performance reasons.