[B] Refactor withReadingGroup HOC
Previously, the withReadingGroups HOC was responsible for fetching a
user's reading groups. As reading groups become more central under the OER
features, and are used outside of the reader, it makes more sense to just
fetch the user's groups when she is authenticated. This allows multiple
instances of the withReadingGroup HOC to exist on a page without each of
them triggering competing fetches.