[F] Source GCS credentials from env, then DB
When Manifold is deployed in a k8s environment, we want to be able to ensure that the
GCS credentials given to it in the environment will always supersede credentials stored
in the database on the Settings object. This change also allows services like sidekiq
to get the correct settings before the database is updated from the environment, which
matters, because these credentails are needed very early in the initialization process,
when the Shrine storage is setup.