[B] Implement Collapse in user notifications
In this instance, the state of the collapsed content is controlled by
a group of radio inputs. So rather than use Collapse.Toggle, I've
retrieved `toggleProps` from the nearest collapse context and applied
"aria-controls" and "aria-expanded" to the inputs. While this isn't
a recognized approach by W3C currently, there is an open issue in the
w3c/aria repo reviewing this design pattern for ARIA 1.3
(https://github.com/w3c/aria/issues/1404). Moreover, GOV.UK (which
opened this issue), a recognized leader in a11y, has found this pattern
to be helpful to users and is already using it. As such, it makes sense
to optimistically implement this with the hope that it will soon
become a recognized pattern. Until that happens, not all browsers
provide the desired feedback to screen reader users. But this seems
limited to Chrome fortunately. Providing an improved experience for
many SR users is preferable to an experience that confuses all SR users.