[C] Remove babel-plugin-i18n-extract
The idea behind this is cool—babel looks through your code and
generates a JSON file of strings from any i18n translation keys you've
declared. In practice, though, this doesn't make sense to use. I think
we want to be more intentional and deliberate in how we structure our
locale strings from the beginning, and this will be more  difficult to
do when we're relying on a plugin to generate those files.