[F] Create EntityCollection components
Creates abstract UI component that basically renders a header, filters,
pagination, and list of entities. This pattern appears repeatedly on the
frontend of the application and has long warranted DRYing out into a
single component. To help with this, this commit also adds a suite of
composed components that shape model-specific data and render an
EntityCollection. These have been applied to projects, project
collections, resources, and resource collections.

The goal is to also use it with collection content blocks and featured
projects, at which point we'll have removed a lot of redundant
components and CSS around rendering lists of entities.

Resolves PROJ-740