[C] Monkey patch Emotion to surpress warning.
This commit introduces a webpack rule that monkey patches an Emotion
warning. We're using Emotion's "advanced" approach to SSR, which means
that the styles are inserted in the head, rather than before each
component. The first child-warning, as we understand it, exists to warn
developers that the insertion of the style tag could change which
element is the first-child, which isn't an issue in our case. Sadly,
emotion doesn't provide a viable way to disable this warning, so we're
monkey patching it here to stop it from spamming the console.

See https://github.com/emotion-js/emotion/issues/1105