[F] Finalize Safari/FF keyboard annotation support
This commit addresses browser inconsistencies around using the
annotation popup that were introduced by 39ae9d0. That commit refactored
the menu using the WAI-ARIA menu design pattern and implemented keyboard
functionality for the first time. However, due to the way Safari and
Firefox interweave the document's focus and selection states (see
https://github.com/ManifoldScholar/manifold/pull/3086#issuecomment-1030141362),
when focus was moved around the menu it triggered a loss of selection
highlighting in these browsers. Because it did this for both pointer and
keyboard interactions, we felt that this was an unacceptable visual
regression, even if it didn't impact how Manifold internally tracks a
selection for annotating, highlighting, sharing, or attaching a resource
to.

As a workaround, this commit tracks whether the user has opened the menu
with a pointer or a keyboard. For pointer users, it now avoids managing
focus, so that Safari and Firefox maintain the current selection. For
keyboard users, focus is managed by the menu, which is necessary for
interaction. This does cause loss of selection highlighting for keyboard
users of these browsers, but this is ultimately unavoidable and beyond
the responsibility of Manifold (see
https://github.com/ManifoldScholar/manifold/issues/2326#issuecomment-523208990
for more on this). This is admittedly not an ideal solution, but it is
the best we can do under the constraints of browser inconsistencies.
Chromium-based browsers continue to work as desired.

Fortunately this is probably a temporary solution, as the CSS Custom
Highlight API should create an alternative, standards-driven path to
storing selection in the browser.

https://www.w3.org/TR/css-highlight-api-1/
https://css-tricks.com/css-custom-highlight-api-early-loo/