[B] Correct behavior when RGs are disabled
The change in 9ccfcb8d811062f43bdcff480161e28c7b70e43c wasn’t a good
solution to the problem described in the commit because it’s difficult
to check settings in the current user middleware. This commit moves the
RG loading into the main Manifold app container and triggers it when a
user logs in, or when the app is mounted.