[B] Fix activeAnnotation prop mismatch
CaptureClick was passing an array of selected annotation IDs up to
Annotatable in the edge case that a clicked annotation contained a link.
Annotatable uses this data to update the active annotation state, but
in all places making use of this state, a string is expected with a
single annotation ID. I've updated the CaptureClick handler to pass
just the first annotation ID in these cases.