[B] Fix annotation popup on iOS
This commit fixes problems with trying to use the reader annotation
popup menu on iOS. Previously when a user tapped a menu item, nothing
happening, as the interpreted mouse event `onClick` wasn't firing in
response to a touch event as expected. I'm wondering whether this
relates to iOS text selection callout also appearing simulatenously, but
I'm not sure. (Also, I looked into hiding that callout, but it's not
possible). In any case, I've worked around this by attaching the same
`onClick` callbacks in the popup menu to `onTouchEnd` events. I've also
added `preventDefault()` for all `onTouchEnd` events so that focus isn't
moved away from the current text selection, which was causing the menu
to become invisible.