[B] Fix occasional YouTube embed failure
It's easy to accidentally pass parameters on the end of an ID when
creating a video resource if you're not versed in URL search params
(e.g. `Zo6UnKr6Bwg&t=21s`). The `&` will cause the embed to fail in the
browser. This commit does a bit of parsing of that ID now to prevent
this.

Resolves #3095
Resolves MNFLD-357