[B] Fix hover, focus styles in Safari
Before 15.4, Safari didn't support the `:focus-visible` pseudo-class.
We've had the `focus-visible` polyfill running, which adds a
`.focus-visible` class that can be used as an alternative style hook,
but we still were using `:focus-visible` a lot in our stylesheets.
We were also occasionally declaring `&:hover, &:focus-visible {}`,
which would cause Safari to skip declarations for both selectors. This
commit replaces all instances of `:focus-visible` with `.focus-visible`
to make these styles work in all browsers we support. In the near
future, we can revert this, but not until <=15.3 Safari is a thing of
the past.