[B] Receive Form.Switch click on input directly
The previous design of this component had the input element hidden, with
the parent label element receiving clicks. For reasons I don't fully
grasp, in the context of the global Drawer component, this caused clicks
to move focus out of the drawer, subsequently causing the drawer's
focus trap to shift focus back into the drawer to the first focusable
element (the close button). Furthermore, this only became an issue when
"tabindex=-1" was added to the document's main element in 04254f8. This
may be related to https://github.com/focus-trap/focus-trap/issues/313.

The easiest solution was just to expand the size of Form.Switch's input
element to fill its parent label. This ensures that clicks hit the input
and keep focus within the trap.

This resolves an issue where clicks