[B] Deactivate focus trap only if close callback exists
Previously the `clickOutsideDeactivates` option was always set to `true`
in the global Drawer component. However, we don't always close the
drawer on clicks outside the trap container, so it doesn't make sense
always activate the trap. Instead, this option is only now set to `true`
if a close callback or close URL is provided.

See also 5cdedb for an earlier attempted fix. This should be more
reliable.

Resolves #3399