[B] Refine key event handling for annotations
Mishandling of keyup and keydown events was causing the following issues
that are now fixed:
* page would scroll when opening annotation with Space key
* annotation drawer would reopen when closed with Enter key

This commit also adjusts the container for capturing selection to allow
tabbing. This helps with an issue in Firefox where selecting text with
a keyboard wasn't working due to caret browsing shifting focus all the
way up to the <body> element and missing the capturing container's key
events. That's not something Manifold can fix, but at least now it's
possible to move focus back to the capturing container.

Resolves MNFLD-393
Resolves MNFLD-401