[B] Correctly set section titles in manifest ingests
Prior to this change, a manifest ingestion would set the section title
to the last value seen in the TOC for that section. This meant that when
a section was referenced multiple times by a nested branch of the TOC,
the most nested value would be assigned to the TOC. With this change,
the first and top-most label in the TOC will be used for the text
section.