[B] Fix GHA apt packages
There is an underlying issue with the cache-apt-pkgs-action we are
using that is yet unresolved, so we are switching to just installing
packages manually with apt.

Additionally, rspec will now bail out early if imagemagick is not
available instead of letting the whole suite run only to fail in 10m.