[C] Improve webpack-dev-server setup
* Resolves an issue where the dev server was started twice, the second
time without HMR, due to an issue with how entries were defined in
browser-dev.config.js. (The previous config wasn't compatible with the
upgraded webpack setup.)
* Removes Node API usage in favor of setting up the dev server via the
config file. This isn't strictly necessary, but it was the simplest way
to resolve the first issue, and simplifies our setup in general.
* Disables liveReload for the dev server, which caused a full page
reload on code changes. Now just HMR is used to show changes.