[B] Properly handle swapping positions in atomic operations
When using drag and drop to reorder reading groups, it is necessary
to disable acts_as_list's functionality conditionally based on whether
or not there are multiple position values in the provided data. If so,
all collectable entries will have acts_as_list temporarily disabled
within the context of the request.

Resolves MNFLD-1007