[B] Fix confirm prop causing server error
`confirm` was passed as a prop when it didn't exist, and so the value
was interpreted as the native `window.confirm()` function. Interestingly,
this led to a `Could not find "store"` error with redux on the server
and a blank screen when visiting the RG edit route on a page reload.