[B] Move focus trap outside of drawer
We've always placed the focus trap just inside the global drawer.
However, because the drawer often has padding, this has caused an issue
where a click on the drawer element's padding would be considered a
click outside the trap, causing the drawer to close.

This commit makes the FocusTrap component the parent of the drawer,
following the recommended pattern in [the library's docs](https://github.com/focus-trap/focus-trap-react).