[B] Fix filter alignment in entity collections
I don't like this solution, but it was the most reasonable way to get
search filters to appear centered with the title in various
EntityCollection patterns on wider container widths. The alternative
was to get rid of flex-wrapping on entity collections and use hard
breakpoints to stack the title and filters, but given how variable
title and search/filter widths are across the patterns, this seemed like
an even worse option.